Web Vitals Tracking Ratio field extracted into separate class
