+++ php/src/class-settings.php
@@ -44,13 +44,6 @@ class Settings {
 	const HARDCODED_TRACKER_CONFIG_FEATURE = 'site_performance_tracker_vitals';
 
 	/**
-	 * Option web_vitals_tracking_ratio name
-	 *
-	 * @var string
-	 */
-	const OPTION_WEB_VITALS_TRACKING_RATIO = 'web_vitals_tracking_ratio';
-
-	/**
 	 * All setting fields
 	 *
 	 * @var array
@@ -67,6 +60,7 @@ class Settings {
 			new MeasurementVersionDimensionField(),
 			new EventMetaDimensionField(),
 			new EventDebugDimensionField(),
+			new WebVitalsTrackingRatioField(),
 		);
 	}
 
@@ -130,14 +124,6 @@ class Settings {
 		foreach ( $this->fields as $field ) {
 			$field->init( $this, self::PAGE_ID, self::SECTION_ID );
 		}
-
-		add_settings_field(
-			self::OPTION_WEB_VITALS_TRACKING_RATIO,
-			__( 'Web Vitals Tracking Ratio', 'site-performance-tracker' ),
-			array( $this, 'web_vitals_tracking_ratio_render' ),
-			self::PAGE_ID,
-			self::SECTION_ID
-		);
 	}
 
 	/**
@@ -145,33 +131,6 @@ class Settings {
 	 */
 	public function settings_section_callback() {
 		echo esc_html( __( 'Update Site Performance Tracker settings', 'site-performance-tracker' ) );
-	}
-
-	/**
-	 * Render Tracking Ratio form input.
-	 */
-	public function web_vitals_tracking_ratio_render() {
-		$options = $this->get_settings();
-		global $tracker_config;
-		$set = false;
-		if ( isset( $tracker_config[ self::OPTION_WEB_VITALS_TRACKING_RATIO ] ) ) {
-			$options[ self::OPTION_WEB_VITALS_TRACKING_RATIO ] = $tracker_config[ self::OPTION_WEB_VITALS_TRACKING_RATIO ];
-			$set                                  = true;
-		}
-		if ( has_filter( 'site_performance_tracker_chance' ) ) {
-			$options[ self::OPTION_WEB_VITALS_TRACKING_RATIO ] = apply_filters( 'site_performance_tracker_chance', 1 );
-			$set                                  = true;
-		}
-		?>
-		<input type='number' name='spt_settings[<?php echo esc_attr( self::OPTION_WEB_VITALS_TRACKING_RATIO ); ?>]' step='0.01' min='0.01' max='1'
-			   value='<?php echo esc_attr( $options[ self::OPTION_WEB_VITALS_TRACKING_RATIO ] ); ?>'
-			   placeholder="Enter between 0 > 1" aria-label="web vitals tracking ratio"
-				<?php if ( $set ) { ?>
-					readonly
-				<?php } ?>>
-		<?php
-
-		$this->show_theme_warning( $set );
 	}
 
 	/**
@@ -222,7 +181,7 @@ class Settings {
 				MeasurementVersionDimensionField::OPTION_MEASUREMENT_VERSION    => '',
 				EventMetaDimensionField::OPTION_EVENT_META                      => '',
 				EventDebugDimensionField::OPTION_EVENT_DEBUG                    => '',
-				self::OPTION_WEB_VITALS_TRACKING_RATIO                          => Plugin::TRACKING_DEFAULT_CHANCE,
+				WebVitalsTrackingRatioField::OPTION_WEB_VITALS_TRACKING_RATIO   => Plugin::TRACKING_DEFAULT_CHANCE,
 			),
 			$options
 		);
