+++ tests/phpunit/class-test-settings.php
@@ -222,7 +222,7 @@ class Test_Settings extends WP_UnitTestCase {
 		$this->assertSame( $field_name, $field['id'] );
 		$this->assertSame( 'Web Vitals Tracking Ratio', $field['title'] );
 
-		$this->assertSame( array( $this->settings, 'web_vitals_tracking_ratio_render' ), $field['callback'] );
+		$this->assertSame( array( $this->settings->fields[5], 'web_vitals_tracking_ratio_render' ), $field['callback'] );
 
 		$this->assertSame( array(), $field['args'] );
 	}
@@ -715,8 +715,10 @@ EOD;
 	}
 
 	public function test_web_vitals_tracking_ratio_render__empty_options() {
+		$this->settings->settings_init();
+
 		ob_start();
-		$this->settings->web_vitals_tracking_ratio_render();
+		$this->settings->fields[5]->web_vitals_tracking_ratio_render();
 		$result = ob_get_contents();
 		ob_end_clean();
 
@@ -731,10 +733,12 @@ EOD;
 	}
 
 	public function test_web_vitals_tracking_ratio_render__set_web_vitals_tracking_ratio() {
+		$this->settings->settings_init();
+
 		add_option( 'spt_settings', array( 'web_vitals_tracking_ratio' => 0.05 ) );
 
 		ob_start();
-		$this->settings->web_vitals_tracking_ratio_render();
+		$this->settings->fields[5]->web_vitals_tracking_ratio_render();
 		$result = ob_get_contents();
 		ob_end_clean();
 
@@ -753,8 +757,10 @@ EOD;
 		global $tracker_config;
 		$tracker_config['web_vitals_tracking_ratio'] = '0.33';
 
+		$this->settings->settings_init();
+
 		ob_start();
-		$this->settings->web_vitals_tracking_ratio_render();
+		$this->settings->fields[5]->web_vitals_tracking_ratio_render();
 		$result = ob_get_contents();
 		ob_end_clean();
 
@@ -771,6 +777,8 @@ EOD;
 	}
 
 	public function test_web_vitals_tracking_ratio_render__apply_filters() {
+		$this->settings->settings_init();
+
 		add_filter(
 			'site_performance_tracker_chance',
 			function () {
@@ -779,7 +787,7 @@ EOD;
 		);
 
 		ob_start();
-		$this->settings->web_vitals_tracking_ratio_render();
+		$this->settings->fields[5]->web_vitals_tracking_ratio_render();
 		$result = ob_get_contents();
 		ob_end_clean();
